requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
async-timeout==4.0.2; python_version >= '3.6'
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
elasticsearch==7.13.0
fastapi-cache2[redis]==0.1.3.4
fastapi==0.63.0
filelock==3.0.12
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0; python_version >= '3.6'
httptools==0.1.1
idna==2.10
importlib-metadata==3.4.0
joblib==1.1.0; python_version >= '3.6'
nltk==3.6.2
pbr==5.5.1
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-json-logger==2.0.1
python-multipart==0.0.5
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
setuptools==60.9.3; python_version >= '3.7'
six==1.16.0
starlette==0.13.6
threadpoolctl==2.1.0
tqdm==4.63.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
wincertstore==0.2
zipp==3.4.0
